moving `pull resource` logic out of pull.js into a sharable location + adding pullAll command which pulls all trns for a project, and cleaning up some spacing -> tabs
